requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
langchain>=0.3.0
langchain-openai>=0.2.0
openai>=1.51.0
pillow==10.2.0
requests==2.31.0
